(WIP) work towards a grid model tuning script.
